# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auto Tournament API",
    "version": "1.0.0",
    "description": "Run a tournament from a bot or a script: teams, players, brackets, matches and results. The dashboard is a client of this API and nothing more. Which game a tournament runs is a module's business, so the endpoints here are the same whether it is Counter-Strike 2 or a game whose captains report their own scores.",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3069",
      "description": "Your MAT instance. Replace with its base URL."
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Service token for machine clients (bots, scripts, CI). Send `Authorization: Bearer <token>`, where the token is one configured in API_TOKENS (full admin) or API_TOKENS_READONLY (GET only). Human admins authenticate with a session cookie instead."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Service token for machine clients (bots, scripts, CI). Send `Authorization: Bearer <token>`, where the token is one configured in API_TOKENS (full admin) or API_TOKENS_READONLY (GET only). Human admins authenticate with a session cookie instead."
      },
      "apiToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Token",
        "description": "Service token, as an alternative to `Authorization: Bearer <token>`."
      },
      "atServerToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auto-Tournament-Token",
        "description": "Server-to-API authentication (webhooks, reports, demo uploads, match config fetch). Send `X-Auto-Tournament-Token: <token>`."
      }
    },
    "schemas": {
      "Server": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique server identifier",
            "example": "cs1"
          },
          "name": {
            "type": "string",
            "description": "Server display name",
            "example": "Server #1"
          },
          "host": {
            "type": "string",
            "description": "Server IP address or hostname",
            "example": "192.168.254.232"
          },
          "port": {
            "type": "integer",
            "description": "Server port",
            "example": 27015
          },
          "password": {
            "type": "string",
            "description": "RCON password",
            "example": "rcon_password"
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the server is enabled",
            "example": true
          },
          "created_at": {
            "type": "integer",
            "description": "Unix timestamp of creation",
            "example": 1699000000
          },
          "updated_at": {
            "type": "integer",
            "description": "Unix timestamp of last update",
            "example": 1699000000
          }
        }
      },
      "CreateServerInput": {
        "type": "object",
        "required": [
          "id",
          "name",
          "host",
          "port",
          "password"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique server identifier",
            "example": "cs1"
          },
          "name": {
            "type": "string",
            "description": "Server display name",
            "example": "Server #1"
          },
          "host": {
            "type": "string",
            "description": "Server IP address or hostname",
            "example": "192.168.254.232"
          },
          "port": {
            "type": "integer",
            "description": "Server port (1-65535)",
            "example": 27015
          },
          "password": {
            "type": "string",
            "description": "RCON password",
            "example": "rcon_password"
          }
        }
      },
      "UpdateServerInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Server display name",
            "example": "Server #1 Updated"
          },
          "host": {
            "type": "string",
            "description": "Server IP address or hostname",
            "example": "192.168.254.232"
          },
          "port": {
            "type": "integer",
            "description": "Server port (1-65535)",
            "example": 27015
          },
          "password": {
            "type": "string",
            "description": "RCON password",
            "example": "new_password"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/events/server/{serverId}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "GET /api/events/server/:serverId",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
...
    "/api/servers": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "List all servers",
        "description": "Get all servers or filter by enabled status",
        "parameters": [
          {
            "in": "query",
            "name": "enabled",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by enabled servers only"
          }
        ],
        "responses": {
          "200": {
            "description": "List of servers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "servers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Server"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Create a new server",
        "description": "Create a new server configuration",
        "parameters": [
          {
            "in": "query",
            "name": "upsert",
            "schema": {
              "type": "boolean"
            },
            "description": "If true, update the server if it already exists"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServerInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Server created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "server": {
                      "$ref": "#/components/schemas/Server"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Server already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
    "/api/servers/batch": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Create multiple servers",
        "description": "Batch create servers. Use ?upsert=true to update existing servers.",
        "parameters": [
          {
            "in": "query",
            "name": "upsert",
            "schema": {
              "type": "boolean"
            },
            "description": "If true, update servers if they already exist"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateServerInput"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Servers created successfully"
          },
          "207": {
            "description": "Partial success (some servers failed)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Servers"
        ],
        "summary": "PATCH /api/servers/batch",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
    "/api/servers/bulk-delete": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "POST /api/servers/bulk-delete",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
    "/api/servers/reset-all-initialization": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "POST /api/servers/reset-all-initialization",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
...
    "/api/servers/{id}": {
      "get": {
        "tags": [
          "Servers"
        ],
        "summary": "Get server by ID",
        "description": "Retrieve a specific server configuration",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Server ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Server found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "server": {
                      "$ref": "#/components/schemas/Server"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "Servers"
        ],
        "summary": "PUT /api/servers/:id",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Servers"
        ],
        "summary": "Update server (partial)",
        "description": "Partially update a server configuration",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server updated"
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Servers"
        ],
        "summary": "Delete server",
        "description": "Remove a server configuration",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server deleted"
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
...
    "/api/servers/{id}/disable": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Disable server",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server disabled"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
    "/api/servers/{id}/enable": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "Enable server",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server enabled"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
    "/api/servers/{id}/reset-initialization": {
      "post": {
        "tags": [
          "Servers"
        ],
        "summary": "POST /api/servers/:id/reset-initialization",
        "description": "Generated from the router. No hand-written `@openapi` block exists for this endpoint yet, so the request and response shapes are not described — read the handler.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiToken": []
          }
        ]
      }
    },
...
  }
}